backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
python-dotenv==1.0.1
Pillow==10.4.0
